% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_age_bin.R
\name{add_age_bin}
\alias{add_age_bin}
\title{add age time bin}
\usage{
add_age_bin(
  data_source,
  age_var_name = "age",
  bin_var_name = "BIN",
  bin_size = 1000,
  sel_method = c("backward", "forward")
)
}
\arguments{
\item{data_source}{Data frame with age represented by \code{age_var_name} variable.}

\item{age_var_name}{Character. Name of the column with age.}

\item{bin_var_name}{Character. Name of the new colun with binned ages.}

\item{bin_size}{Numeric. Size of the time bin.}

\item{sel_method}{Character. Which direction should the values be binned.
\itemize{
\item \code{"backward"} - c(59, 105) -> c(0, 100)
\item \code{"forward"} - c(59, 105) -> c(100, 200)
}}
}
\description{
add age time bin
}
