% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data_partition.R
\name{add_data_partition}
\alias{add_data_partition}
\title{Add partition of data}
\usage{
add_data_partition(data_source_var, data_source_levels, age_var = "age")
}
\arguments{
\item{data_source_var}{Data frame with the variables to partition}

\item{data_source_levels}{Data frame with ages of levels}

\item{age_var}{Name (in quotes) of variable which define the age of samples}
}
\description{
Add partition for each column for chronologically ordered data using regression trees
}
