% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca_execute_canoco.r
\name{dcca_execute_canoco}
\alias{dcca_execute_canoco}
\title{Execute CANOCO}
\usage{
dcca_execute_canoco(
  data_resp,
  data_pred,
  default_name = "xx",
  downweight = FALSE
)
}
\arguments{
\item{data_resp}{data frame with response data (pollen counts or percentages)}

\item{data_pred}{data frame with required predictors (all will be used)}

\item{default_name}{how to name CON, OUT and SOL files (e.g. xx.con, xx.out, xx.sol)
Default value "xx"
The `default_name`` is also used for produced data files:
\itemize{
\item xx-pred.dta
\item xx-resp.dta
}}

\item{downweight}{logical - whether to downweight rare species
Default value FALSE}
}
\value{
\itemize{
\item \code{eig} - numeric vector - eigenvalues for the first four axes
\item \code{tot_inertia} - total variation in (transformed) response data
\item \code{turn} - numeric vector with turnover values
\item \code{case_e} - numeric matrix with CaseE scores for (up to) first 3 axes
\item \code{case_r} - numeric matrix with CaseR scores for first 4 axes
}
}
\description{
Execute CANOCO
}
\section{Performs single execution of command-line CANOCO program}{
with specified response and predictor data, using DCCA after
square-root transforming. Returns analysis results as
a list (described below). On failures calls the stop
function, so it might be wise to enclose the call in a try
construct to catch such cases
}

\author{
Petr Smilauer
}
