% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca_make_con_file.R
\name{dcca_make_con_file}
\alias{dcca_make_con_file}
\title{Cretae CON file}
\usage{
dcca_make_con_file(
  con_file_path,
  resp_file_path,
  pred_file_path,
  default_name = "xx",
  downweight = FALSE,
  n_segments = 26,
  n_rescale = 4,
  rescale_threshold = 0
)
}
\arguments{
\item{con_file_path}{qualified name of CON file to create}

\item{resp_file_path}{name of (CANOCO-format) file with response data}

\item{pred_file_path}{name of (CANOCO-format) file with explanatory data}

\item{default_name}{temporary file name default_name (for .OUT and .SOL files)
(default "xx")}

\item{downweight}{logical value: down-weighting of rare species
(default \code{FALSE})}

\item{n_segments}{number of segments to use for detrending
(default 26)}

\item{n_rescale}{number of times to apply non-linear rescaling
(default 4)}

\item{rescale_threshold}{threshold for non-linear rescaling
(default 0.0)}
}
\value{
TRUE on success, FALSE otherwise
}
\description{
Creates CANOCO 4.5x project (.CON) file for a specific
type of analysis (DCCA with detrending by segments)
}
\author{
Petr Smilauer
}
