% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_estimate.R
\name{diversity_estimate}
\alias{diversity_estimate}
\title{Estimation of interpolation and extrapolation of individual-based Hill number}
\usage{
diversity_estimate(
  data_source,
  sel_method = c("taxonomic", "phylogenetic_diversity", "phylogenetic_nri",
    "phylogenetic_nti"),
  sample_size,
  rand = 999,
  iterations = 1000,
  abundance_weighted = TRUE,
  phylogenetic_tree = NULL
)
}
\arguments{
\item{data_source}{Data frame with rows as levels and columns as taxa. First
row should be named \code{sample_id}.}

\item{sel_method}{Selected method of diversify estimation c("taxonomic_randomised", "taxonomic_rarefy")}

\item{sample_size}{minimum sample size}

\item{rand}{number of randomisations}
}
\value{
Tibble including phylogenetic diversity metric for each level.
}
\description{
Estimation of interpolation and extrapolation of individual-based Hill number
}
