% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_estimate_functional.R
\name{diversity_estimate_functional}
\alias{diversity_estimate_functional}
\title{Estimate of functional diversity}
\usage{
diversity_estimate_functional(
  data_source_community,
  data_source_traits,
  sel_method = c("rao", "mpd", "simpson"),
  rand = 1000
)
}
\arguments{
\item{data_source_community}{Data matrix. Rows as levels and columns as data_source_community. Row names
should be the \code{sample_id}.}

\item{data_source_traits}{Data matrix. Row names and col names must be the same
as data_source_community in \code{data_source_community}}

\item{rand}{Number of randomization}
}
\value{
TODO
}
\description{
TODO
}
