% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_estimate_phylogenetic.R
\name{diversity_estimate_phylogenetic}
\alias{diversity_estimate_phylogenetic}
\title{Estimating phylogenetic diversity}
\usage{
diversity_estimate_phylogenetic(
  data_matrix,
  tree,
  dataset_id,
  sel_method = c("diversity", "NRI", "NTI"),
  abundance_weighted = TRUE,
  rand = 999,
  iterations = 1000
)
}
\arguments{
\item{data_matrix}{Data frame with rows as levels and columns as taxa}

\item{tree}{Phylogenetic backbone tree constructed using ape package}

\item{sel_method}{Selected method for diversity estimation.
\itemize{
\item \code{"diversity"} - Faith's phylogenetic diversity
\item \code{"NRI"} = Net Relatedness Index
\item \code{"NTI"} = Nearest Taxon Index
}}

\item{abundance_weighted}{Logical. It FALSE, presence absence data will
be used.}

\item{rand}{Number of randomization}

\item{iterations}{Number of iterations to use for each randomization
(for independent swap and trial null models)}
}
\value{
Data frame with diversity metric estimated for each level (sample).
Possible outputs depending on the \code{sel_method}:
\itemize{
\item \code{n_taxa} -
\item \code{faith_pd_diversity} -
\item \code{mean_pairwise_distances} -
\item \code{net_relatedness_index} -
\item \code{mean_nearest_taxon_distances} -
\item \code{nearest_taxon_index} -
\item \code{z_score} -
\item \code{p_value} -
}
}
\description{
Function for estimating one of three metrics of phylogenetic
diversity
}
