% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_estimate_randomise.R
\name{diversity_estimate_randomise}
\alias{diversity_estimate_randomise}
\title{Estimating diversity and evenness based on resampling without replacement}
\usage{
diversity_estimate_randomise(
  data_matrix,
  sample_size,
  set_margin = c(1, 2),
  rand = 999
)
}
\arguments{
\item{data_matrix}{Data matrix}

\item{sample_size}{The minimum sample size}

\item{set_margin}{Are levels as columns or rows? 1 = rows, 2 = columns}

\item{rand}{Number of randomisations}
}
\value{
Data frame with  Hill numbers N0, N1, N2, the evenness ratios N2/N1, N2/N0, and the
modified versions of N2/N1 and N2/N0. In addition, it estimate unbiased confidence intervals
}
\description{
Estimating diversity and evenness based on resampling without replacement. Code adapted from Chao et al. 2014
}
