% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_estimate_rarefy.R
\name{diversity_estimate_rarefy}
\alias{diversity_estimate_rarefy}
\title{Estimation of interpolation and extrapolation of individual-based Hill number}
\usage{
diversity_estimate_rarefy(data_matrix, sample_size, set_margin = 1)
}
\arguments{
\item{data_matrix}{Data matrix}

\item{sample_size}{minimum sample size}

\item{set_margin}{Are levels as columns or rows? 1 = rows, 2 = columns}
}
\value{
Matrix with Hills number N0, N1, and N2, evenness ratios, and error estimates
}
\description{
Estimation of interpolation and extrapolation of individual-based Hill number
}
