% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_custom_gam.R
\name{fit_custom_gam}
\alias{fit_custom_gam}
\title{Fit custom single term GAM model}
\usage{
fit_custom_gam(
  x_var = "age",
  y_var = "var",
  error_family = "gaussian(link = 'identity')",
  smooth_basis = c("tp", "cr"),
  data_source,
  weights_var = NULL,
  sel_k = 10
)
}
\arguments{
\item{x_var}{Character. Name of the X-variable}

\item{y_var}{Character. Name of the Y-variable}

\item{error_family}{Character. Name of the error-family to be used}

\item{smooth_basis}{Character. Name of the Smooth basis to use}

\item{data_source}{Data.frame}

\item{weights_var}{Character. Name of the variable to use as weights}

\item{sel_k}{Define \code{k} (wiggliness)}
}
\value{
Fitted GAM model
}
\description{
A wrapper function for \code{mgcv::gam} to help fit GAM models
functionally.
}
