% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gam_safely.R
\name{fit_gam_safely}
\alias{fit_gam_safely}
\title{Safely fit single term GAM}
\usage{
fit_gam_safely(
  data_source,
  x_var = "age",
  y_var = "var",
  smooth_basis = c("tp", "cr"),
  error_family = "gaussian(link = 'identity')",
  sel_k = 10,
  max_k = 10
)
}
\arguments{
\item{data_source}{Data.frame}

\item{x_var}{Character. Name of the X-variable}

\item{y_var}{Character. Name of the Y-variable}

\item{smooth_basis}{Character. Name of the Smooth basis to use}

\item{error_family}{Character. Name of the error-family to be used}

\item{sel_k}{Preferred \code{k} (wiggliness)}

\item{max_k}{Maximum \code{k} which can be used}
}
\description{
A wrapper function for \code{fit_custom_gam}. The \code{k}is compared to
the number of samples and to the \code{max_k}. If there is a zero variability in
the term, \code{lm} is fitted instead. If no model was cretaed (due to errors),
the function retun \code{NA_real}
}
