% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_multiple_gams.R
\name{fit_multiple_gams}
\alias{fit_multiple_gams}
\title{Fit multiple individual generalized additive models (GAMs)}
\usage{
fit_multiple_gams(
  data_source,
  x_var = "age",
  y_var = "var",
  group_var = "dataset_id",
  smooth_basis = c("tp", "cr"),
  error_family = "gaussian(link = 'identity')",
  weights_var = NULL,
  max_k = 10,
  max_iterations = 200,
  verbose = TRUE
)
}
\arguments{
\item{data_source}{Data frame with \code{x_var}, \code{y_var}, and \code{group_var}}

\item{x_var}{Character. Name of the X-variable}

\item{y_var}{Character. Name of the Y-variable}

\item{group_var}{Character. Name of the variable used for identification of
groups.}

\item{smooth_basis}{Character. Name of the Smooth basis to use}

\item{error_family}{Character. Name of the error-family to be used}

\item{weights_var}{Character. Name of the variable to use as weights}

\item{max_k}{Maximum \code{k} which can be used}

\item{max_iterations}{Numeric. Maximum number of iteration for GAM to try.}

\item{verbose}{Logical. Should additional messages be output?}
}
\value{
A data frame containing the original grouping variable and the
corresponding fitted GAM model for each group.
}
\description{
This function fits multiple generalized additive models (GAMs) using the
\code{fit_gam_safely()} function. GAMs are fitted
to each dataset defined by the \code{group_var} argument in the input
\code{data_source} data frame. The number of models fitted is equal to the
number of unique levels of the \code{group_var}. The function returns a data
frame with one row per group, including the original group identifier and the
fitted GAM model for that group.
}
\seealso{
\code{\link[=fit_gam_safely]{fit_gam_safely()}}
}
