% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ordination_dca.R
\name{fit_ordination_dca}
\alias{fit_ordination_dca}
\title{Fit DCA}
\usage{
fit_ordination_dca(data_source)
}
\arguments{
\item{data_source}{data frame with response data (pollen counts or percentages). Sample ID as
rownames.}
}
\value{
\itemize{
\item \code{ordination} - ordination object of class "decorana"
\item \code{case_r} - numeric matrix with CaseR scores for first 4 axes
}
}
\description{
Fits a DCA using \code{\link[vegan:decorana]{vegan::decorana()}}.
}
\seealso{
\code{\link[vegan:decorana]{vegan::decorana()}}
}
