% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_change_points.R
\name{get_change_points}
\alias{get_change_points}
\title{Detect change points in a vector}
\usage{
get_change_points(data_source, direction = c("front", "back"))
}
\arguments{
\item{data_source}{Vector with values (numeric, character, or factor)}

\item{direction}{Which direction should the values be compared ("front" or "back")}
}
\value{
Vector with binary values of length = \code{data_source} + 1,
where \code{1} represents a change point and \code{0} represents no change point.
}
\description{
This function returns a binary vector indicating where a
change occurs in the input numeric, character, or factor vector.
The change points are identified based on the direction parameter provided.
}
