% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_change_points_all.R
\name{get_change_points_all}
\alias{get_change_points_all}
\title{Detect change points for all columns}
\usage{
get_change_points_all(
  data_source,
  var_name_order = "age",
  use_inbetween_values = TRUE,
  sel_output = c("only_inbetween", "both"),
  var_name_detection = NULL,
  direction = c("front", "back")
)
}
\arguments{
\item{data_source}{Data.frame with columns to detect change points in}

\item{var_name_order}{Character.
Name of the columns, which should be use as a order of values}

\item{use_inbetween_values}{Logical. If \code{TRUE} additional values will be placed
in between values of \code{var_name_order}.
#' @param sel_output
Character. Only used for \code{use_inbetween_values} == \code{TRUE}
Define what should the function output.
\itemize{
\item \code{"only_inbetween"} - only inbetween values,
\item \code{"both"} - both inbetween values and original data
}}

\item{var_name_detection}{Character. Optional. Columns to use to change point detection
will be automatically selected if they partly matched \code{var_name_detection}.
For all columns use `NULL``.}

\item{direction}{Which direction should the values be compared ("front" or "back")}
}
\value{
Data.frame with \code{var_name_order} and all columns detected based on
\code{var_name_detection}, with \code{1} represent change points.
}
\description{
Detect change points for all columns
}
