% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_density.R
\name{get_density}
\alias{get_density}
\title{Get density from vector of numbers}
\usage{
get_density(data_source, values_range, reflected = TRUE, ...)
}
\arguments{
\item{data_source}{Vector with numeric values}

\item{values_range}{Specific know range of the \code{data_source}}

\item{reflected}{Logical. Should reflecting technique be used to fix
the boundary effect? (from Density Estimation for Statistics and Data
Analysis by B.W.Silverman)}

\item{...}{Additional parameters used in \code{stats::density}}
}
\value{
Data.frame with the \code{data_source} values and estimated density
}
\description{
Get density from vector of numbers
}
