% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.data.distribution.by.var.R
\name{map.data.distribution.by.var}
\alias{map.data.distribution.by.var}
\title{Plot geographic distribution}
\usage{
map.data.distribution.by.var(
  data_source,
  var,
  title = var,
  coord_long,
  coord_lat,
  custom_palette,
  palette_set = "Set1"
)
}
\arguments{
\item{data_source}{Data.frame with data}

\item{var}{Character. Name of the variable which should be used to
colour-group data}

\item{title}{Optional. Name of the variable to display}

\item{coord_long}{Optional. Vector containing min and max longitude
to display}

\item{coord_lat}{Optional. Vector containing min and max latitude
to display}

\item{custom_palette}{Optional. Named vector containing colour for levels
of \code{var}}

\item{palette_set}{if \code{custom_palette} is not specified, function will
create new palette using \code{RColorBrewer::brewer.pal}. \code{palette_set} can
specify which general palette to use.}
}
\value{
ggplot with the geographic distribution of data
}
\description{
Plot geographic distribution
}
