% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melodic.R
\name{melodic}
\alias{melodic}
\title{Mean Dissimilarity Components}
\usage{
melodic(samp, dis, type = "both")
}
\arguments{
\item{samp}{community matrix; sites in lines, species in columns}

\item{dis}{dissimilarity matrix}

\item{type}{\itemize{
\item \code{"both"} - for results with abundances weighted and non-weighted
\item \code{"abundance"} - for results with abundances weighted
\item \code{"presence"} - for results with abundances non-weighted
}}
}
\description{
Function \code{melodic} obtained from https://doi.org/10.1007/s00442-016-3546-0
}
\author{
Carlos P. Carmona
}
