% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_regression_partition.R
\name{mv_regression_partition}
\alias{mv_regression_partition}
\title{Multivariate regression trees}
\usage{
mv_regression_partition(
  data_source_counts,
  data_source_levels,
  rand = 1000,
  transformation = c("chisq", "hellinger", "none")
)
}
\arguments{
\item{data_source_counts}{Data frame with pollen data. Each row represent one
level (sample) and each column represent one taxon. Table must contain
\code{sample_id} column with unique values corresponding to \code{data_source_levels}}

\item{data_source_levels}{Data frame with level ages, with each row represent
one level (sample). Table must contain two columns: \code{sample_id} with unique
values corresponding to \code{data_source_levels}, and \code{age} with ages of each
level}

\item{rand}{Number of multiple cross-validations}

\item{transformation}{Selection of data transformation options. \code{hellinger} =
Hellinger, \code{chisq} = Chisq, \code{none} = without transformation. Both Hellinger
and Chisq are recommend for proportional/percentage data}
}
\description{
Transform pollen data into percentages, apply Chisq
tranformation, and then apply multivariate regression trees.
}
