% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temporal_trend.R
\name{plot_temporal_trend}
\alias{plot_temporal_trend}
\title{Plot temporal trend}
\usage{
plot_temporal_trend(
  data_source,
  x_var = "age",
  x_var_name = "Time (ka BP)",
  y_var = "var",
  y_var_name = "Estimate",
  sel_type = c("single", "group"),
  group_var = NULL,
  display_error = FALSE,
  y_limits = NULL,
  sel_y_trans = "identity",
  sel_x_trans = "reverse",
  x_ticks = seq(0, 12000, 4000),
  y_ticks = seq(0, 1, 0.5),
  n_y_ticks = 3,
  show_rmse = TRUE,
  show_bin_summary = TRUE,
  summary_bin_size = 1000,
  deafult_color_line = "black",
  default_color_rmse_highlight = "red",
  default_color_summary_line = "blue",
  default_color_summary_fill = "lightblue",
  group_color_pallete = NULL,
  line_size = 1,
  line_alpha = 1,
  error_alpha = 0.2,
  summary_line_size = 0.5,
  summary_alpha = 0.5,
  default_text_size = 16
)
}
\arguments{
\item{data_source}{DESCRIPTION.}

\item{x_var}{Character. Name of the X-variable}

\item{x_var_name}{Character. label of the X-variable}

\item{y_var}{Character. Name of the Y-variable}

\item{y_var_name}{Character. label of the Y-variable}

\item{sel_type}{Character. The type of time series.
\itemize{
\item \code{"single"} - only one series
\item \code{"group"} - several series
}}

\item{group_var}{Character. Name of the variable defining the groups
in \code{sel_type}.}

\item{display_error}{Logical. Should error bar be displayed?}

\item{y_limits}{Either \code{NULL} or 2 value vector to set limits.}

\item{sel_y_trans}{Character. Name of tranfsormation used in \code{scale_y_continuous}}

\item{sel_x_trans}{Character. Name of tranfsormation used in \code{scale_x_continuous}}

\item{x_ticks}{Numeric vector with x-ticks.}

\item{y_ticks}{Numeric vector with y-ticks or \code{"auto"} to set them automaticaly}

\item{n_y_ticks}{If \code{y_ticks} = \code{"auto"}, then set number of automatic ticks}

\item{show_rmse}{Locigal. Should RMSE visualisation be plotted.
Only works for \code{group_var} = \code{"group"}}

\item{show_bin_summary}{Locigal. Should age bin summary be plotted.
Only works for \code{group_var} = \code{"group"}.}

\item{summary_bin_size}{Size of time bins to use.
Only works for \code{show_bin_summary} = \code{TRUE}.}

\item{deafult_color_line}{Character with HEX code.}

\item{default_color_rmse_highlight}{Character with HEX code.}

\item{default_color_summary_line}{haracter with HEX code.}

\item{default_color_summary_fill}{haracter with HEX code.}

\item{group_color_pallete}{Optional. Name vector with colors for levels of \code{group_var}}

\item{line_size}{Numeric.}

\item{line_alpha}{Numeric.}

\item{error_alpha}{Numeric.}

\item{summary_line_size}{Numeric.}

\item{summary_alpha}{Numeric.}

\item{default_text_size}{Numeric.}
}
\description{
Plot temporal trend
}
