% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REcopol-package.R
\docType{package}
\name{REcopol-package}
\alias{REcopol}
\alias{REcopol-package}
\title{REcopol: Analyse fossil pollen records}
\description{
The package assemble function used in community ecology and present them in easy-to-use wrapper functions.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
