% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_partition.R
\name{regression_partition}
\alias{regression_partition}
\title{Partition data using regression tree}
\usage{
regression_partition(data_source, var = "", age_var = "")
}
\arguments{
\item{data_source}{Data frame with the variables to partition}

\item{var}{Character. Name (in quotes) of variable to partition}

\item{age_var}{Character. Name (in quotes) of variable which define the age
of samples}
}
\value{
\itemize{
\item \code{rpart_result} - the regression tree generated by the \code{mvpart::rpart()}
\item \code{pruned_tree} - the pruned version of the regression tree
\item \code{rpart_groups} - the number of segments generated by the regression
partitioning analysis
\item \code{rpart_change_points} - the change points in the independent variable
(age) that correspond to the partitioning
\item \code{rpart_partitions} -  the partitioned data, with an additional column
indicating the segment to which each row belongs
}
}
\description{
Partition chronologically ordered data using regression trees.
This function performs regression partitioning analysis on a given data
source to identify segments with different relationships between the
dependent variable and a specified independent variable. The output includes
the partitioned data, the resulting regression tree, the number of resulting
segments, and the change points in the independent variable age that
correspond to the partitioning.
}
