% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renumber_groups.R
\name{renumber_groups}
\alias{renumber_groups}
\title{Re-number cluster groups}
\usage{
renumber_groups(clusgr)
}
\arguments{
\item{clusgr}{Vector with integers}
}
\description{
Utility function to re-number vector of integers to simple
sequence increased by one
}
