% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tranform_percentage_data.R
\name{tranform_percentage_data}
\alias{tranform_percentage_data}
\title{Transform pollen data using selected transformation}
\usage{
tranform_percentage_data(
  data_source,
  transformation = c("chisq", "hellinger", "none")
)
}
\arguments{
\item{data_source}{Data frame with pollen data. Each row represent one
level (sample) and each column represent one taxon. Table must contain
\code{sample_id} column with unique values.}

\item{transformation}{Selection of data transformation options. \code{hellinger} =
Hellinger, \code{chisq} = Chisq, \code{none} = without transformation. Both Hellinger
and Chisq are recommend for proportional/percentage data}
}
\description{
Transform pollen data using one of the selected transformation
functions
}
