% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_into_proportions.R
\name{transfer_into_proportions}
\alias{transfer_into_proportions}
\title{Transform pollen data into percentages}
\usage{
transfer_into_proportions(
  data_source,
  method = c("percentages", "proportions")
)
}
\arguments{
\item{data_source}{Data frame with pollen data. Each row represent one
level (sample) and each column represent one taxon. Table must contain
\code{sample_id} column with unique values.}

\item{method}{variable to select result as either percentage (\code{percentages})
or proportions (\code{percentages}).}
}
\description{
Tranform pollen data into percentages (or proportions)
}
