% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca_utility_funcions.R
\name{util_clean_files}
\alias{util_clean_files}
\title{Delete files within folder}
\usage{
util_clean_files(sel_folder_path, file_to_omit = "canoco.exe")
}
\description{
Try to delete all files within folder but \code{file_to_omit}
}
\keyword{internal}
