% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_inbetween_values.R
\name{util_get_inbetween_values}
\alias{util_get_inbetween_values}
\title{Get in-between values}
\usage{
util_get_inbetween_values(data_source)
}
\arguments{
\item{data_source}{Vector with values}
}
\description{
Add values in-between each pair or values in a vector. Position
of each value is calculated as average of 2 original values.
}
