% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca_utility_funcions.R
\name{util_get_line_by_tabs}
\alias{util_get_line_by_tabs}
\title{Parse line by tabs}
\usage{
util_get_line_by_tabs(sel_line, id_from, id_to)
}
\description{
Takes a character string and parses it with TAB separator,
returning specified range of string entries
}
\author{
Petr Smilauer
}
