% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca_utility_funcions.R
\name{util_read_lines_until}
\alias{util_read_lines_until}
\title{Read file until found selected text}
\usage{
util_read_lines_until(connection_file, sel_text)
}
\description{
Takes an open connection to a text file and reads from it individual
lines until it arrives to a line starting with specified text
Found line is returned.
Passed connection (connection_file) must be already open for reading
The function returns empty string to indicate a failure
}
\author{
Petr Smilauer
}
\keyword{internal}
