% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca_utility_funcions.R
\name{util_read_n_lines}
\alias{util_read_n_lines}
\title{Read number of lines}
\usage{
util_read_n_lines(connection_file, num_lines)
}
\description{
Reads specified number of lines from the passed
connection. Last read line is returned
}
\author{
Petr Smilauer
}
