% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_renumber_groups.R
\name{util_renumber_groups}
\alias{util_renumber_groups}
\title{Re-number cluster groups}
\usage{
util_renumber_groups(clusgr)
}
\arguments{
\item{clusgr}{Vector with integers}
}
\description{
Utility function to re-number vector of integers to simple
sequence of numbers
}
