% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data_partition.R
\name{add_data_partition}
\alias{add_data_partition}
\title{Add partition of data}
\usage{
add_data_partition(data_source_var, data_source_levels, age_var = "age")
}
\arguments{
\item{data_source_var}{A dataframe with samples as rows and variables as
columns. This dataframe must contain a "sample_id" column.}

\item{data_source_levels}{A dataframe with one row per sample, where the
sample_id column must match the data_source_var dataframe. This dataframe
must contain an "age" column.}

\item{age_var}{A character string indicating the name of the age column in
the data_source_levels dataframe. Default value is "age".}
}
\value{
A dataframe with the same columns as data_source_var plus a column
for each partition, named as variable name plus "_part".
}
\description{
Add partition for each column for chronologically ordered
data using regression trees.
}
