% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcca_make_can_file.R
\name{dcca_make_can_file}
\alias{dcca_make_can_file}
\title{Export CAN file}
\usage{
dcca_make_can_file(sel_path, sel_data)
}
\value{
FALSE on failure, TRUE otherwise
}
\description{
Saves a data frame into a file in Canoco free format
path is the fully qualified name of the file to create
data is the data frame
}
\author{
Petr Smilauer
}
