% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_estimate.R
\name{diversity_estimate}
\alias{diversity_estimate}
\title{Estimation of interpolation and extrapolation of individual-based Hill number}
\usage{
diversity_estimate(
  data_source,
  method = c("rarefy", "randomised"),
  sample_size,
  rand = 999
)
}
\arguments{
\item{data_source}{Data frame with rows as levels and columns as taxa}

\item{method}{Selected method of diversify estimation c("randomised", "rarefy")}

\item{sample_size}{minimum sample size}

\item{rand}{number of randomisations}
}
\value{
Tibble including phylogenetic diversity metric for each level.
}
\description{
Estimation of interpolation and extrapolation of individual-based Hill number
}
