% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_estimate_functional.R
\name{diversity_estimate_functional}
\alias{diversity_estimate_functional}
\title{Estimate functional diversity}
\usage{
diversity_estimate_functional(
  data_source_community,
  data_source_traits,
  sel_method = c("rao", "mpd", "simpson"),
  abundance_weighted = TRUE,
  rand = 1000
)
}
\arguments{
\item{data_source_community}{Data matrix. Rows as levels and columns as taxa.
Row names should be the \code{sample_id}.}

\item{data_source_traits}{Data frame. Row names be the same taxa as
in \code{data_source_community} and vice versa.}

\item{sel_method}{Selected method for diversity estimation:
\itemize{
\item \code{"rao"} - Rao's quadratic diversity
\item \code{"mpd"} = Mean Pairwise Distance
\item \code{"simpson"} = Simpsons Diversity Index
}}

\item{abundance_weighted}{Logical. It FALSE, presence absence data will be used.}

\item{rand}{Numeric. Number of randomization}
}
\value{
Data frame with diversity metric estimated for each level (sample).
Possible outputs depending on the \code{sel_method}:
\itemize{
\item \code{fd_rao} - Rao's quadratic diversity
\item \code{fd_mpd} - Mean Pairwise Distance
\item \code{fd_simpson} - Simpsons Diversity Index
\item \code{z_score} - Standarise Effect Size (SES), calculated as
(observed value - mean randomised value ) / sd of randomised value
}
}
\description{
TODO
}
\author{
Triin Reitalu, Ondrej Mottl
}
