% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_estimate_phylogenetic.R
\name{diversity_estimate_phylogenetic}
\alias{diversity_estimate_phylogenetic}
\title{Estimating phylogenetic diversity}
\usage{
diversity_estimate_phylogenetic(
  data_source,
  tree,
  dataset_id,
  method = c("diversity", "NRI", "NTI"),
  abundance_weighted = TRUE,
  rand = 999,
  iterations = 1000
)
}
\arguments{
\item{data_source}{Data frame with rows as levels and columns as taxa}

\item{tree}{Phylogenetic diversity tree constructed using ape package}

\item{dataset_id}{Optional. unique dataset identification as character}

\item{method}{Selected method for diversity estimation.
"diversity" = Faith's phylogenetic diversity
"NRI" = Net Relatedness Index
"NTI" = Nearest Taxon Index}

\item{abundance_weighted}{Logical. It FALSE, presence absence data will
be used.}

\item{rand}{Number of randomization}

\item{iterations}{Number of iterations to use for each randomization
(for independent swap and trial null models)}
}
\value{
Tibble with Hill numbers N0, N1, N2, the evenness ratios N2/N1, N2/N0, and the
modified versions of N2/N1 and N2/N0
}
\description{
Function for estimating one of three metrics of phylogenetic
diversity
}
