% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_estimate_phylogenetic.R
\name{diversity_estimate_phylogenetic}
\alias{diversity_estimate_phylogenetic}
\title{Estimating phylogenetic diversity}
\usage{
diversity_estimate_phylogenetic(
  data_matrix,
  phylogenetic_tree,
  sel_method = c("diversity", "nri", "nti"),
  abundance_weighted = TRUE,
  rand = 999,
  iterations = 1000
)
}
\arguments{
\item{data_matrix}{Data matrix. Rows as levels and columns as data.
Row names should be the \code{sample_id}.}

\item{phylogenetic_tree}{Phylogenetic backbone phylogenetic_tree constructed using \code{ape} package}

\item{sel_method}{Selected method for diversity estimation.
\itemize{
\item \code{"diversity"} - Faith's phylogenetic diversity
\item \code{"nri"} = Net Relatedness Index
\item \code{"nti"} = Nearest Taxon Index
}}

\item{abundance_weighted}{Logical. It FALSE, presence absence data will be used.}

\item{rand}{Numeric. Number of randomization}

\item{iterations}{Number of iterations to use for each randomization
(for independent swap and trial null models)}
}
\value{
Data frame with diversity metric estimated for each level (sample).
Possible outputs depending on the \code{sel_method}:
\itemize{
\item \code{n_taxa} - Taxonomic richness
\item \code{pd_faith} - Faith's phylogenetic diversity
\item \code{pd_mpd} - Mean Pairwise Distance
\item \code{pd_nri} - Net Relatedness Index
\item \code{pd_mntd} - Mean Nearest Taxon Distance
\item \code{pd_nti} - Nearest Taxon Index
\item \code{z_score} - Standarise Effect Size (SES), calculated as
(observed value - mean randomised value ) / sd of randomised value
}
}
\description{
Function for estimating one of three metrics of phylogenetic diversity
}
