% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_estimate_taxonomic.R
\name{diversity_estimate_taxonomic}
\alias{diversity_estimate_taxonomic}
\title{Estimation taxonomic diversity}
\usage{
diversity_estimate_taxonomic(data_matrix, sample_size)
}
\arguments{
\item{data_matrix}{Data matrix. Rows as levels and columns as taxa.
Row names should be the \code{sample_id}.}

\item{sample_size}{Numeric. Minimum sample size}
}
\value{
#' Data frame with diversity metric estimated for each level (sample):
\itemize{
\item \code{n0} - Hill' N0
\item \code{n1} - Hill' N1
\item \code{n2} - Hill' N2
\item \code{n1_minus_n2} - evenness ratios (N1-N2)
\item \code{n2_divided_by_n1} - evenness ratios (N2/N1)
\item \code{n1_divided_by_n0} - evenness ratios (N1/N0)
}
}
\description{
Estimation of interpolation and extrapolation of individual-based
Hill numbers
}
