% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_custom_gam.R
\name{fit_custom_gam}
\alias{fit_custom_gam}
\title{Fit custom single term GAM model}
\usage{
fit_custom_gam(
  y_var,
  x_var,
  error_family,
  smooth_basis = c("tp", "cr"),
  data_source,
  weights_var = NULL,
  sel_k = 10
)
}
\arguments{
\item{y_var}{Character. Name of the Y-variable}

\item{x_var}{Character. Name of the X-variable}

\item{error_family}{Character. Name of the error-family to be used}

\item{smooth_basis}{Character. Name of the Smooth basis to use}

\item{data_source}{Data.frame}

\item{weights_var}{Character. Name of the variable to use as weights}

\item{sel_k}{Define \code{k} (wiggliness)}
}
\value{
Fitted GAM model
}
\description{
Fit custom single term GAM model
}
