% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hgam.R
\name{fit_hgam}
\alias{fit_hgam}
\title{Fit single term hierarchical GAM model}
\usage{
fit_hgam(
  x_var = "age",
  y_var = "var",
  group_var = "dataset_id",
  error_family = "gaussian(link = 'identity')",
  smooth_basis = c("tp", "cr"),
  data_source,
  sel_k = 10,
  sel_m = NULL,
  common_trend = TRUE,
  use_parallel = TRUE,
  use_discrete = FALSE,
  max_itiration = 200
)
}
\arguments{
\item{x_var}{Character. Name of the X-variable}

\item{y_var}{Character. Name of the Y-variable}

\item{group_var}{Character. Name of grouping variable}

\item{error_family}{Character. Name of the error-family to be used}

\item{smooth_basis}{Character. Name of the smooth basis to use for \code{x_var}}

\item{data_source}{Data.frame with columns whose names are set by \code{y_var},
\code{x_var}, \code{group_var}}

\item{sel_k}{Numeric. Define \code{k} (wiggliness) for \code{x_var}}

\item{sel_m}{Numeric. User specify the order of the penalty for this term.
if \code{NULL}, function will use \code{1} or \code{2} depending on the presence of common
trend.}

\item{common_trend}{Logical. Should hGAM have a common shared trend?}

\item{use_parallel}{Logical. Should computation use parallel?}

\item{use_discrete}{Logical. Should \code{discrete} agument be used for parallel
computation?}

\item{max_itiration}{Numeric. Maximum number of iteration for hGAM to try.}
}
\value{
Fitted hGAM model
}
\description{
Fit a hierarchical GAM model with/without a single common
smoother (\code{common_trend}) plus group-level smoothers with differing
wiggliness (random effect). If there is less number of groups in the dataset
than 2, normal GAM model will be fitted.
If \code{use_parallel} is \code{TRUE}, number of cores is automatically detected.
}
\seealso{
{mgcv::bam()}
}
