% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_multiple_gams.R
\name{fit_multiple_gams}
\alias{fit_multiple_gams}
\title{Fit multiple individual GAM models}
\usage{
fit_multiple_gams(
  data_source,
  x_var = "age",
  y_var = "var",
  group_var = "dataset_id",
  smooth_basis = c("tp", "cr"),
  error_family = "gaussian(link = 'identity')",
  max_k = 10
)
}
\arguments{
\item{data_source}{Data frame with \code{x_var}, \code{y_var}, and \code{group_var}}

\item{x_var}{Character. Name of the X-variable}

\item{y_var}{Character. Name of the Y-variable}

\item{smooth_basis}{Character. Name of the Smooth basis to use}

\item{error_family}{Character. Name of the error-family to be used}

\item{max_k}{Maximum \code{k} which can be used}
}
\description{
Helper function to apply multiple\code{fit_gam_safely}.
}
\seealso{
\code{\link[=fit_gam_safely]{fit_gam_safely()}}
}
