% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ordination.R
\name{fit_ordination}
\alias{fit_ordination}
\title{Fit ordination}
\usage{
fit_ordination(
  data_source_community,
  data_source_predictors = NULL,
  var_name_pred = c("age"),
  var_name_cond = NULL,
  sel_method = c("cca", "rda", "dca"),
  transform_to_percentage = FALSE,
  tranformation = c("none", "chisq", "hellinger")
)
}
\arguments{
\item{data_source_community}{Data.frame with community data. Each row is sample. First columns is
\code{sample_id}, each other is taxa.}

\item{data_source_predictors}{Data.frame with predictors. First columns is \code{sample_id}. Other columns
can be predictors and/or conditional variables.}

\item{var_name_pred}{Character. Vector with the names of predicotr variables.}

\item{var_name_cond}{Character. Optional name of conditional variable}

\item{sel_method}{\itemize{
\item \code{"cca"} - Correspondence Analysis
\item \code{"rda"} - Redundancy Analysis
\item \code{"dca"} - Detrended Correspondence Analysis
}}

\item{transform_to_percentage}{Logical. Should community data be tranformed into percentages?}

\item{tranformation}{Logical. Which tranformation should be applied?
\itemize{
\item "none" -  without transformation
\item "chisq" - Chisq tranformation
\item "hellinger" - Hellinger tranformation
}
It is recomneded to apply tranformation to percentage data.}
}
\value{
Ordniation from \code{vegan} package
}
\description{
This is a wrapper function to perform standard data preparation
and either unconstrained (DCA) or (partialy-)consatrined ordination (CCA/RDA)
with fossil pollen data. By default constrained ordination is constrained by
\code{"age"}.
}
