% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ordination_dcca.R
\name{fit_ordination_dcca}
\alias{fit_ordination_dcca}
\title{Fit DCCA using CANOCO}
\usage{
fit_ordination_dcca(
  data_source_resp,
  data_source_pred,
  sel_complexity = c("linear", "poly_2", "poly_3", "auto"),
  downweight = FALSE
)
}
\arguments{
\item{data_source_resp}{data frame with response data (pollen counts or percentages). Sample ID as
rownames.}

\item{data_source_pred}{Data frame with the predictor values (depth or age).  Sample ID as
rownames.
The number of rows must be identical with the number of rows in \code{data_source_resp}}

\item{sel_complexity}{Indicate how the predictor should be used
/itemize{
/item \verb{"linear"`` - a linear predictor /item }"poly_2"\verb{ - a polynomial predictor of second order /item `"poly_3"} - a polynomial predictor of third order
/item `"auto"`` - a complexity will be selected based on the shape of DCA
}}

\item{downweight}{logical, whether to downweight rare species or not}
}
\value{
/itemize{
/item \code{eig} - numeric vector - eigenvalues for the first four axes
/item  \code{tot_inertia} - total variation in (transformed) response data
/item \code{turn} - numeric vector with turnover values
/item \code{case_r} - numeric matrix with CaseE scores for (up to) first 3 axes
/item \code{case_r} - numeric matrix with CaseR scores for first 4 axes
/item \code{sel_complexity} - see \code{sel_complexity} agument description
}
}
\description{
Fits a DCCA using CANOCO.exe program
}
