% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_change_points.R
\name{get_change_points}
\alias{get_change_points}
\title{Detect change points in a vector}
\usage{
get_change_points(data_source, direction = c("front", "back"))
}
\arguments{
\item{data_source}{Vector with values (numeric, character, or factor)}

\item{direction}{Which direction should the values be compared ("front" or "back")}
}
\value{
Vector with binary values of length = \code{data_source} + 1.
\code{1} represent change point
}
\description{
Detect change points in a vector
}
