% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melodic.R
\name{melodic}
\alias{melodic}
\title{Mean Dissimilarity Components}
\usage{
melodic(data_source_community, data_source_dissimilarity)
}
\arguments{
\item{data_source_community}{community matrix; sites in lines, species in columns}

\item{data_source_dissimilarity}{dissimilarity matrix}
}
\description{
Function \code{melodic} adapted from
https://doi.org/10.1007/s00442-016-3546-0
}
\author{
Ondrej Mottl, Carlos P. Carmona
}
