% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temporal_trend.R
\name{plot_temporal_trend}
\alias{plot_temporal_trend}
\title{Plot temporal trend}
\usage{
plot_temporal_trend(
  data_source,
  x_var = "age",
  x_var_name = "Time (ka BP)",
  y_var = "var",
  y_var_name = "Estimate",
  sel_type = c("single", "group"),
  group_var = NULL,
  display_error = FALSE,
  y_limits = NULL,
  sel_y_trans = "identity",
  sel_x_trans = "reverse",
  x_ticks = seq(0, 12000, 4000),
  y_ticks = seq(0, 1, 0.5),
  n_y_ticks = 3,
  show_rmse = TRUE,
  show_bin_summary = TRUE,
  summary_bin_size = 1000,
  deafult_color_line = "black",
  default_color_rmse_highlight = "red",
  default_color_summary_line = "blue",
  default_color_summary_fill = "lightblue",
  group_color_pallete = NULL,
  line_size = 1,
  line_alpha = 1,
  error_alpha = 0.2,
  summary_line_size = 0.5,
  summary_alpha = 0.5,
  default_text_size = 16
)
}
\arguments{
\item{data_source}{data.frame to be plotted.}

\item{x_var}{Character. The name of the column in \code{data_source} containing
the x-variable (default: "age")}

\item{x_var_name}{Character. The name of the x-axis variable
(default: "Time (ka BP)")}

\item{y_var}{Character. The name of the column in \code{data_source} containing
the y-variable}

\item{y_var_name}{Character. The name of the y-axis variable}

\item{sel_type}{Character. The type of time series.
\itemize{
\item \code{"single"} - only one series
\item \code{"group"} - several series
}
The function will plot a single line with error bars (if \code{display_error} is
\code{TRUE}). If \code{"group"}, the function will plot multiple lines with different
colours based on the `group_var``}

\item{group_var}{Character. Name of the variable defining the groups
in \code{sel_type}.}

\item{display_error}{Logical. Should an error bar be displayed?}

\item{y_limits}{The limits of the y-axis in the form of a vector of length 2
(default: \code{NULL}). If \code{NULL}, the function will calculate the limits
automatically.}

\item{sel_y_trans}{Character. Name of tranfsormation used in \code{scale_y_continuous}
(default: \code{"identity"})}

\item{sel_x_trans}{Character. Name of tranfsormation used in \code{scale_x_continuous}
(default: \code{"reverse"})}

\item{x_ticks}{Numeric vector with x-ticks.}

\item{y_ticks}{Numeric vector with y-ticks or \code{"auto"} to set them automatically}

\item{n_y_ticks}{If \code{y_ticks} = \code{"auto"}, then set number of automatic ticks}

\item{show_rmse}{Locigal. Should root mean square error (RMSE) visualisation
be plotted.
Only works for \code{group_var} = \code{"group"}}

\item{show_bin_summary}{Locigal. Should an age bin summary be plotted?
Only works for \code{group_var} = \code{"group"}.}

\item{summary_bin_size}{Size of time bins to use.
Only works for \code{show_bin_summary} = \code{TRUE}.}

\item{deafult_color_line}{Character with HEX code.}

\item{default_color_rmse_highlight}{Character with HEX code.}

\item{default_color_summary_line}{Character with HEX code.}

\item{default_color_summary_fill}{Character with HEX code.}

\item{group_color_pallete}{Optional. Name vector with colours for levels of \code{group_var}}

\item{line_size}{Numeric.}

\item{line_alpha}{Numeric.}

\item{error_alpha}{Numeric.}

\item{summary_line_size}{Numeric.}

\item{summary_alpha}{Numeric.}

\item{default_text_size}{Numeric.}
}
\description{
a customizable plotting function that takes in several arguments to plot
a temporal trend of a variable over time. The function allows the user to
customize the x and y variables, the display of error bars, the grouping
of the data, the colour and size of the lines, and many other graphical
features. The function returns a plot object that can be further
customized or saved. This function is currently \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
