% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_partition.R
\name{regression_partition}
\alias{regression_partition}
\title{Partition data in groups}
\usage{
regression_partition(data_source, var = "", age_var = "")
}
\arguments{
\item{data_source}{Data frame with the variables to partition}

\item{var}{Name (in quotes) of variable to partition}

\item{age_var}{Name (in quotes) of variable which define the age os samples}
}
\description{
Partition chronologically ordered data using regression trees
}
